feat: AttendanceJob 설정
